add mongodb support, create db and handlers folders
