+++ db/Tasks.go
@@ -1,0 +1,7 @@
+package db
+
+type Task struct {
+	ID      int    `json:"ID"`
+	Name    string `json:"name"`
+	Content string `json:"content"`
+}
