+++ go.mod
@@ -1,0 +1,8 @@
+module github.com/Sirpyerre/taskList
+
+go 1.16
+
+require (
+	github.com/gorilla/mux v1.8.0
+	go.mongodb.org/mongo-driver v1.8.3
+)
