+++ handlers/CreateHandle.go
@@ -1,0 +1,26 @@
+package handlers
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+
+	"github.com/Sirpyerre/taskList/db"
+)
+
+func createTask(w http.ResponseWriter, r *http.Request) {
+	var newTask db.Task
+	reqBody, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		fmt.Fprintf(w, "Insert a Valid task")
+	}
+
+	json.Unmarshal(reqBody, &newTask)
+	newTask.ID = len(tasks) + 1
+	tasks = append(tasks, newTask)
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusCreated)
+	json.NewEncoder(w).Encode(newTask)
+}
